% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mtCompare}
\alias{mtCompare}
\title{Compare package status on two different MRAN snapshot dates.}
\usage{
mtCompare(dates, type = "source")
}
\arguments{
\item{dates}{Character vector with two ISO dates, e.g. \code{c("2014-10-01", "2014-11-01")}.}

\item{type}{character, indicating the type of package to download and install.  Possible values are (currently) "source", "mac.binary", "mac.binary.mavericks" and "win.binary".}
}
\description{
Compare package status on two different MRAN snapshot dates.
}
\examples{
# Summarise package differences between two snapshot dates
x <- mtCompare(c("2014-10-01", "2014-11-15"))
str(x)
}

