% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setdiffinv}
\alias{setdiffinv}
\title{Inverse setdiff.}
\usage{
setdiffinv(x, y)
}
\arguments{
\item{x}{vector (of the same mode) containing a sequence of items (conceptually) with no duplicated values.}

\item{y}{vector (of the same mode) containing a sequence of items (conceptually) with no duplicated values.}
}
\description{
Wrapper around \code{\link[base]{setdiff}} to invert the arguments.
}

