% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setSnapshot}
\alias{setSnapshot}
\title{Set MRAN snapshot date.}
\usage{
setSnapshot(date, MRANroot = "http://mran.revolutionanalytics.com/snapshot/")
}
\arguments{
\item{date}{Character vector with single ISO date, e.g. "2014-09-17".}

\item{MRANroot}{URL containing snapshots}
}
\description{
Set MRAN snapshot date.
}
\examples{
oldRepo <- getOption("repos")

# Empty date field returns current repo
setSnapshot()
setSnapshot(NULL)

# Valid snapshot date
setSnapshot("2014-11-16")

# Invalid snapshot date (in future), returns error
try(
  setSnapshot("2024-11-16")
)

options(repos = oldRepo)
}

